+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `TERR_Res` in `tega/err.h` for error handling.
+
 #### `tega/fundamental.h`
 
 Interface for general programming utilities.